implemnt admin and views

+++ teamup/src/main/java/it/fgm/teamup/model/Teammate.java
@@ -3,7 +3,7 @@ package it.fgm.teamup.model;
 public class Teammate extends Utente {
     private int id;
 
-    public long getId() { return id; }
+    public int getId() { return id; }
 
     public void setId(int id) { this.id = id; }
 }
